Boating to Save the Ocean
Our project is Boating to Save the Ocean!!! Boating to Save the Ocean is a game that helps teach the importance of keeping the ocean clean. In the game, the player controls the boat to pick up the trash in a limited amount of time. If they succeed, the help keep the ocean clean. If they fail, it will display the consequences of polluting the ocean with trash.